// packages/core/src/models/enums/Race.ts
/**
 * This file contains code derived from or adapted from:
 * Eternaltwin DinoRPG
 * Upstream file: https://gitlab.com/eternaltwin/dinorpg/dinorpg/-/blob/3a73bbc6d751e4916cc5fd2e5f23bc2cfd42fc6d/core/src/models/enums/RaceEnum.mts
 *
 * Copyright in the original contributions remains with the respective
 * authors and contributors.
 *
 * Modified by DinoRPG Remastered contributors on 2026-01-20.
 * See NOTICE.md and the Git history for provenance and modification details.
 *
 * SPDX-License-Identifier: AGPL-3.0-or-later
 */
export enum RaceEnum {
	MOUEFFE = 1,
	MOUEFFE_DEMON = 2,
	PIGMOU = 3,
	PIGMOU_DEMON = 4,
	WINKS = 5,
	WINKS_DEMON = 6,
	PLANAILLE = 7,
	PLANAILLE_DEMON = 8,
	CASTIVORE = 9,
	ROCKY = 10,
	PTEROZ = 11,
	NUAGOZ = 12,
	SIRAIN = 13,
	HIPPOCLAMP = 14,
	GORILLOZ = 15,
	GORILLOZ_DEMON = 16,
	WANWAN = 17,
	WANWAN_DEMON = 18,
	SANTAZ = 19,
	FEROSS = 20,
	KABUKI = 21,
	KABUKI_DEMON = 22,
	MAHAMUTI = 23,
	SOUFFLET = 24,
	TOUFUFU = 25,
	QUETZU = 26,
	SMOG = 27,
	TRICERAGNON = 28
}